+++ src/datumaro/cli/commands/require_project/versioning/commit.py
@@ -3,18 +3,21 @@
 # SPDX-License-Identifier: MIT
 
 import argparse
+import logging as log
 
 from datumaro.util.scope import scope_add, scoped
 
 from ....util import MultilineFormatter
 from ....util.project import load_project
 
+deprecated = "[DEPRECATED, will be removed in 1.12]"
+
 
 def build_parser(parser_ctor=argparse.ArgumentParser):
     parser = parser_ctor(
-        help="Create a revision",
-        description="""
-        Creates a new revision from the current state of the working directory.|n
+        help=f"{deprecated} Create a revision",
+        description=f"""
+        {deprecated} Creates a new revision from the current state of the working directory.|n
         |n
         Examples:|n
         - Create a revision:|n
@@ -62,6 +65,7 @@ def get_sensitive_args():
 
 @scoped
 def commit_command(args):
+    log.warning("This command is deprecated and will be removed in Datumaro 1.12")
     project = scope_add(load_project(args.project_dir))
 
     old_tree = project.head